+++ src/components/main/DisplaySection.tsx
