+++ src/components/main/FriendSelection.tsx
