+++ src/types/orderForm.ts
@@ -1,0 +1,4 @@
+export type FormValues = {
+  message: string;
+  sender: string;
+};
